Clean implementation even more

+++ build.sh
@@ -12,23 +12,48 @@ function custom_wait()
 
 function rm_build_dirs()
 {
-    rm -rf "$proj_dir/host_side/build"
-    rm -rf "$proj_dir/pico_side/build"
+    if [ -d "$proj_dir/host_side/build" ]; then
+        echo "Removing host build directory"
+        rm -rf "$proj_dir/host_side/build"
+    fi
+    if [ -d "$proj_dir/pico_side/build" ]; then
+        echo "Removing pico build directory"
+        rm -rf "$proj_dir/pico_side/build"
+    fi
     return 0
 }
 
 function configure()
 {
-    mkdir -p "$proj_dir/host_side/build"
-    cd "$proj_dir/host_side/build"
-    cmake -DCMAKE_BUILD_TYPE=Release -DPICO_BOARD=pico_w ..
-    custom_wait
-
-    mkdir -p "$proj_dir/pico_side/build"
-    cd "$proj_dir/pico_side/build"
-    cmake -DCMAKE_BUILD_TYPE=Release -DPICO_BOARD=pico_w ..
-    custom_wait
-
+    if [ $# -eq 0 ]; then
+        echo "Configuring all"
+        mkdir -p "$proj_dir/host_side/build"
+        cd "$proj_dir/host_side/build"
+        cmake -DCMAKE_BUILD_TYPE=Release -DPICO_BOARD=pico_w ..
+        custom_wait
+    elif [ $# -eq 1 ]; then
+        if [ "$1" == "host" ]; then
+            echo "Configuring host"
+            mkdir -p "$proj_dir/host_side/build"
+            cd "$proj_dir/host_side/build"
+            cmake -DCMAKE_BUILD_TYPE=Release -DPICO_BOARD=pico_w ..
+            custom_wait
+        elif [ "$1" == "pico" ]; then
+            echo "Configuring pico"
+            mkdir -p "$proj_dir/pico_side/build"
+            cd "$proj_dir/pico_side/build"
+            cmake -DCMAKE_BUILD_TYPE=Release -DPICO_BOARD=pico_w ..
+            custom_wait
+        else
+            echo "Invalid argument"
+            cd "$proj_dir"
+            return 1
+        fi
+    else
+        echo "Invalid number of arguments"
+        cd "$proj_dir"
+        return 1
+    fi
 
     cd "$proj_dir"
 
@@ -37,13 +62,35 @@ function configure()
 
 function build()
 {
-    cd "$proj_dir/host_side/build"
-    make -j$(nproc) all
-    custom_wait
-
-    cd "$proj_dir/pico_side/build"
-    make -j$(nproc) all
-    custom_wait
+    if [ $# -eq 0 ]; then
+        echo "Building all"
+        cd "$proj_dir/host_side/build"
+        make all
+        custom_wait
+        cd "$proj_dir/pico_side/build"
+        make all
+        custom_wait
+    elif [ $# -eq 1 ]; then
+        if [ "$1" == "host" ]; then
+            echo "Building host"
+            cd "$proj_dir/host_side/build"
+            make all
+            custom_wait
+        elif [ "$1" == "pico" ]; then
+            echo "Building pico"
+            cd "$proj_dir/pico_side/build"
+            make all
+            custom_wait
+        else
+            echo "Invalid argument"
+            cd "$proj_dir"
+            return 1
+        fi
+    else
+        echo "Invalid number of arguments"
+        cd "$proj_dir"
+        return 1
+    fi
 
     cd "$proj_dir"
 
@@ -53,9 +100,28 @@ function build()
 function main()
 {
     rm_build_dirs
-    configure
-    build
+    if [ $# -eq 0 ]; then
+        echo "Processing all"
+        configure
+        build
+    elif [ $# -eq 1 ]; then
+        if [ "$1" == "host" ]; then
+            echo "Processing host"
+            configure host
+            build host
+        elif [ "$1" == "pico" ]; then
+            echo "Processing pico"
+            configure pico
+            build pico
+        else
+            echo "Invalid argument"
+            return 1
+        fi
+    else
+        echo "Invalid number of arguments"
+        return 1
+    fi
     return 0
 }
 
-main
+main "$@"
